M5/Portafolio/requirements.txt
requests
beautifulsoup4
python-nmap
python-docx
pdfkit
difflib
